Add files via upload
The Road Traffic Accident Dataset contains information about road traffic accidents. It includes data on the total number of vehicles involved, the month and year of the accidents, the total number of accidents, the number of fatal and non-fatal accidents, the number of people killed, and the number of people injured.